+++ src/hangar/repository.py
@@ -10,7 +10,7 @@ from .context import Environments
 from .diagnostics import graphing, ecosystem
 from .records import heads, parsing, summarize, vcompat
 from .checkout import ReaderCheckout, WriterCheckout
-from .utils import is_valid_directory_path, is_suitable_user_key
+from .utils import is_valid_directory_path, is_suitable_user_key, is_ascii
 
 
 class Repository(object):
@@ -447,15 +447,32 @@ class Repository(object):
 
         return commit_hash
 
-    def create_branch(self, name: str, base_commit: str = None) -> str:
+    def create_branch(self, name: str, base_commit: str = None) -> heads.BranchHead:
         """create a branch with the provided name from a certain commit.
 
         If no base commit hash is specified, the current writer branch ``HEAD``
-        commit is used as the base_commit hash for the branch. Note that
+        commit is used as the ``base_commit`` hash for the branch. Note that
         creating a branch does not actually create a checkout object for
         interaction with the data. to interact you must use the repository
         checkout method to properly initialize a read (or write) enabled
         checkout object.
+
+            >>> from hangar import Repository
+            >>> repo = Repository('foo/path/to/dir')
+
+            >>> repo.create_branch('testbranch')
+                BranchHead(name='testbranch', digest='b66b...a8cc')
+            >>> repo.list_branches()
+                ['master', 'testbranch']
+            >>> co = repo.checkout(write=True, branch='testbranch')
+            >>> co.metadata['foo'] = 'bar'
+            >>> # add data ...
+            >>> newDigest = co.commit('added some stuff')
+
+            >>> repo.create_branch('new-changes', base_commit=newDigest)
+                BranchHead(name='new-changes', digest='35kd...3254')
+            >>> repo.list_branches()
+                ['master', 'new-changes', 'testbranch']
 
         Parameters
         ----------
@@ -468,33 +485,158 @@ class Repository(object):
 
         Returns
         -------
-        str
-            name of the branch which was created
+        heads.BranchHead
+            NamedTuple[str, str] with fields for `name` and `digest` of the branch
+            created (if the operation was successful)
+
+        Raises
+        ------
+        ValueError
+            If the branch name provided contains characters outside of alpha-numeric
+            ascii characters and ".", "_", "-" (no whitespace).
+        ValueError
+            If the branch already exists.
+        RuntimeError
+            If the repository does not have at-least one commit on the `default`
+            (ie. `master`) branch.
         """
         self.__verify_repo_initialized()
-        if not is_suitable_user_key(name):
-            e = ValueError(
-                f'Branch name provided: `{name}` invalid. Must only contain '
-                f'alpha-numeric or "." "_" "-" ascii characters.')
+        if (not is_ascii(name)) or (not is_suitable_user_key(name)):
+            e = ValueError(f'Branch name provided: {name} invalid. Must contain '
+                           f'only alpha-numeric or "." "_" "-" ascii characters.')
             raise e from None
-        didCreateBranch = heads.create_branch(
+        createdBranch = heads.create_branch(
             branchenv=self._env.branchenv,
             name=name,
             base_commit=base_commit)
-        return didCreateBranch
+        return createdBranch
 
-    def remove_branch(self, name):  # pragma: no cover
-        """Not Implemented
+    def remove_branch(self, name: str, *, force_delete: bool = False) -> heads.BranchHead:
+        """Permanently delete a branch pointer from the repository history.
+
+        Since a branch (by definition) is the name associated with the HEAD
+        commit of a historical path, the default behavior of this method is to
+        throw an exception (no-op) should the ``HEAD`` not be referenced as an
+        ancestor (or at least as a twin) of a separate branch which is
+        currently *ALIVE*. If referenced in another branch's history, we are
+        assured that all changes have been merged and recorded, and that this
+        pointer can be safely deleted without risk of damage to historical
+        provenance or (eventual) loss to garbage collection.
+
+            >>> from hangar import Repository
+            >>> repo = Repository('foo/path/to/dir')
+
+            >>> repo.create_branch('first-testbranch')
+            BranchHead(name='first-testbranch', digest='9785...56da')
+            >>> repo.create_branch('second-testbranch')
+            BranchHead(name='second-testbranch', digest='9785...56da')
+            >>> repo.list_branches()
+            ['master', 'first-testbranch', 'second-testbranch']
+            >>> # Make a commit to advance a branch
+            >>> co = repo.checkout(write=True, branch='first-testbranch')
+            >>> co.metadata['foo'] = 'bar'
+            >>> co.commit('added some stuff')
+            '3l253la5hna3k3a553256nak35hq5q534kq35532'
+            >>> co.close()
+
+            >>> repo.remove_branch('second-testbranch')
+            BranchHead(name='second-testbranch', digest='9785...56da')
+
+        A user may manually specify to delete an un-merged branch, in which
+        case the ``force_delete`` keyword-only argument should be set to
+        ``True``.
+
+            >>> # check out master and try to remove 'first-testbranch'
+            >>> co = repo.checkout(write=True, branch='master')
+            >>> co.close()
+
+            >>> repo.remove_branch('first-testbranch')
+            Traceback (most recent call last):
+                ...
+            RuntimeError: ("The branch first-testbranch is not fully merged. "
+            "If you are sure you want to delete it, re-run with "
+            "force-remove parameter set.")
+            >>> # Now set the `force_delete` parameter
+            >>> repo.remove_branch('first-testbranch', force_delete=True)
+            BranchHead(name='first-testbranch', digest='9785...56da')
+
+        It is important to note that *while this method will handle all safety
+        checks, argument validation, and performs the operation to permanently
+        delete a branch name/digest pointer, **no commit refs along the history
+        will be deleted from the Hangar database**.* Most of the history contains
+        commit refs which must be safe in other branch histories, and recent
+        commits may have been used as the base for some new history. As such, even
+        if some of the latest commits leading up to a deleted branch ``HEAD`` are
+        orphaned (unreachable), the records (and all data added in those commits)
+        will remain on the disk.
+
+        In the future, we intend to implement a garbage collector which will remove
+        orphan commits which have not been modified for some set amount of time
+        (probably on the order of a few months), but this is not implemented at the
+        moment.
+
+        Should an accidental forced branch deletion occur, *it is possible to
+        recover* and create a new branch head pointing to the same commit. If
+        the commit digest of the removed branch ``HEAD`` is known, its as simple as
+        specifying a name and the ``base_digest`` in the normal
+        :meth:`create_branch` method. If the digest is unknown, it will be a
+        bit more work, but some of the developer facing introspection tools /
+        routines could be used to either manually or (with minimal effort)
+        programmatically find the orphan commit candidates. If you find
+        yourself having accidentally deleted a branch, and must get it back,
+        please reach out on the `Github Issues
+        <https://github.com/tensorwerk/hangar-py/issues>`__ page. We'll gladly
+        explain more in depth and walk you through the process in any way we
+        can help!
+
+        Parameters
+        ----------
+        name : str
+            name of the branch which should be deleted. This branch must exist, and
+            cannot refer to a remote tracked branch (ie. origin/devbranch), please
+            see exception descriptions for other parameters determining validity of
+            argument
+        force_delete : bool, optional
+            If True, remove the branch pointer even if the changes are un-merged in
+            other branch histories. May result in orphaned commits which may be
+            time-consuming to recover if needed, by default False
+
+        Returns
+        -------
+        :class:`~heads.BranchHead`
+            NamedTuple[str, str] with fields for `name` and `digest` of the branch
+            pointer deleted.
+
+        Raises
+        ------
+        ValueError
+            If a branch with the provided name does not exist locally
+        PermissionError
+            If removal of the branch would result in a repository with zero local
+            branches.
+        PermissionError
+            If a write enabled checkout is holding the writer-lock at time of this
+            call.
+        PermissionError
+            If the branch to be removed was the last used in a write-enabled
+            checkout, and whose contents form the base of the staging area.
+        RuntimeError
+            If the branch has not been fully merged into other branch histories,
+            and ``force_delete`` option is not ``True``.
         """
         self.__verify_repo_initialized()
-        raise NotImplementedError()
+        res = heads.remove_branch(branchenv=self._env.branchenv,
+                                  refenv=self._env.refenv,
+                                  name=name,
+                                  force_delete=force_delete)
+        return res
 
     def list_branches(self) -> List[str]:
         """list all branch names created in the repository.
 
         Returns
         -------
-        list of str
+        List[str]
             the branch names recorded in the repository
         """
         self.__verify_repo_initialized()